Moved exception to DI namespace.
